requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
ibm-cloud-sdk-core==3.20.6
ibm-watson==8.1.0
idna==3.8
Janome==0.5.0
# logging==0.4.9.6
lxml==5.3.0
numpy==2.1.1
plac==1.4.3
portalocker==2.10.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
sacrebleu==2.4.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sumeval==0.2.2
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websocket-client==1.8.0
